server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
ultralytics==8.3.0
python-multipart==0.0.9
Pillow==10.4.0
